@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.6.5+13
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -39,7 +39,7 @@ dependencies:
   provider: ^6.0.1
 
   # navigator
-  auto_route: ^5.0.2
+  auto_route: ^6.0.0+1
 
   # web navigator for SSO
   webview_flutter: ^4.0.1
@@ -76,7 +76,7 @@ dependencies:
 
   # Notifications
   flutter_local_notifications: ^12.0.2
-  unifiedpush: ^4.0.0
+  unifiedpush: ^5.0.0
   desktop_notifications: ^0.6.3
 
   # Flushbar
@@ -108,12 +108,13 @@ dependencies:
   flutter_datetime_picker: ^1.5.1
   intl: ^0.18.0
   scroll_to_index: ^3.0.1
+  video_player: ^2.5.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   hive_generator: ^2.0.0
-  auto_route_generator: ^5.0.2
+  auto_route_generator: ^6.0.0
   build_runner: ^2.1.2
   import_sorter: ^4.6.0
   flutter_lints: ^2.0.1
