@@ -22,6 +22,7 @@ import '../../partials/feed/topic_list_tile.dart';
 import '../../partials/post/post_writer_modal.dart';
 import '../../router.gr.dart';
 
+@RoutePage()
 class CalendarEventPage extends StatefulWidget {
   final Room room;
   const CalendarEventPage({Key? key, required this.room}) : super(key: key);
@@ -101,17 +102,15 @@ class CalendarEventPageState extends State<CalendarEventPage> {
                 builder: (context, snapshot) {
                   return LayoutView(
                       controller: controller,
-                      customHeader: CustomHeader(
-                        title: "Event",
-                        actionButton: [
-                          IconButton(
-                              icon: const Icon(Icons.edit),
-                              onPressed: () {
-                                context
-                                    .pushRoute(SocialSettingsRoute(room: room));
-                              }),
-                        ],
-                      ),
+                      customHeaderText: "Event",
+                      customHeaderActionsButtons: [
+                        IconButton(
+                            icon: const Icon(Icons.edit),
+                            onPressed: () {
+                              context
+                                  .pushRoute(SocialSettingsRoute(room: room));
+                            }),
+                      ],
                       room: room,
                       headerHeight: 280,
                       mainWidth: double.infinity,
@@ -265,9 +264,7 @@ class CalendarEventPageState extends State<CalendarEventPage> {
                                 ),
                               ),
                             ],
-                          ),
-                      headerChildBuilder: ({required bool displaySideBar}) =>
-                          Container());
+                          ));
                 });
           }),
     );
