@@ -3,15 +3,15 @@ import 'package:flutter/material.dart';
 import 'package:matrix/matrix.dart';
 import 'package:minestrix/partials/components/minestrix/minestrix_title.dart';
 import 'package:minestrix/utils/minestrix/minestrix_community_extension.dart';
-import 'package:minestrix_chat/minestrix_chat.dart';
 import 'package:minestrix_chat/partials/matrix/matrix_image_avatar.dart';
 import 'package:minestrix_chat/utils/matrix_widget.dart';
 
 import '../../../partials/components/layouts/custom_header.dart';
-import '../../../partials/navigation/rightbar.dart';
 import '../../../partials/minestrix_room_tile.dart';
+import '../../../partials/navigation/rightbar.dart';
 import '../../../router.gr.dart';
 
+@RoutePage()
 class CommunityPage extends StatefulWidget {
   const CommunityPage({Key? key}) : super(key: key);
 
