@@ -42,12 +42,41 @@ class FollowingIndicator extends StatelessWidget {
     if (room.space == null && room.room == null) return Container();
 
     bool following = room.room != null;
+    if (!following) {
+      return CustomFutureButton(
+          expanded: false,
+          padding: const EdgeInsets.all(6),
+          icon: Icon(following ? Icons.person : Icons.person_add),
+          onPressed: () async => follow(Matrix.of(context).client, context),
+          children: [Text(following ? "Following" : "Follow")]);
+    }
 
-    return CustomFutureButton(
-        expanded: false,
-        padding: const EdgeInsets.all(6),
-        icon: Icon(following ? Icons.person : Icons.person_add),
-        onPressed: () async => follow(Matrix.of(context).client, context),
-        children: [Text(following ? "Following" : "Follow")]);
+    return PopupMenuButton(
+        onSelected: (value) async {
+          await follow(Matrix.of(context).client, context);
+        },
+        itemBuilder: (context) => [
+              const PopupMenuItem(
+                  value: "unfollow",
+                  child: ListTile(
+                      leading: Icon(
+                        Icons.person_remove,
+                        color: Colors.red,
+                      ),
+                      title: Text('Unfollow',
+                          style: TextStyle(color: Colors.red)))),
+            ],
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Row(
+            children: [
+              Icon(following ? Icons.person : Icons.person_add),
+              const SizedBox(
+                width: 5,
+              ),
+              const Text("Following"),
+            ],
+          ),
+        ));
   }
 }
