@@ -9,6 +9,7 @@ import 'package:settings_ui/settings_ui.dart';
 import '../../partials/components/layouts/custom_header.dart';
 import '../../router.gr.dart';
 
+@RoutePage()
 class SettingsStorysPage extends StatefulWidget {
   const SettingsStorysPage({Key? key}) : super(key: key);
 
@@ -39,8 +40,8 @@ class _SettingsStorysPageState extends State<SettingsStorysPage> {
                     title: Text(room.getLocalizedDisplayname(
                         const MatrixDefaultLocalizations())),
                     value: Text("${room.summary.mJoinedMemberCount} followers"),
-                    onPressed: (context) =>
-                        context.navigateTo(const SettingsAccountRoute()),
+                    onPressed: (context) => context
+                        .navigateTo(SettingsStorysDetailRoute(room: room)),
                   ),
               ],
             ),
