@@ -1,14 +1,12 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:matrix/matrix.dart';
-import 'package:minestrix_chat/partials/matrix/matrix_image_avatar.dart';
-import 'package:minestrix_chat/utils/matrix_widget.dart';
-
-import 'package:provider/provider.dart';
-
 import 'package:minestrix/partials/components/layouts/custom_header.dart';
+import 'package:minestrix_chat/utils/matrix_widget.dart';
+import 'package:provider/provider.dart';
 
 import '../../utils/managers/theme_manager.dart';
 
+@RoutePage()
 class SettingsThemePage extends StatefulWidget {
   const SettingsThemePage({Key? key}) : super(key: key);
 
